HNSW-E2E: 1. Moved resources to dedicated sub-module.
