+++ iris-mpc-upgrade-hawk/tests/assets/node-config/docker/node-0-genesis-0.toml
