+++ iris-mpc-upgrade-hawk/tests/assets/node-config/docker/node-1-genesis-0.toml
