+++ iris-mpc-upgrade-hawk/tests/resources/mod.rs
@@ -1,0 +1,11 @@
+mod iris_codes;
+mod iris_deletions;
+mod iris_modifications;
+mod iris_shares;
+mod node_config;
+mod utils;
+
+pub(crate) use iris_deletions::read_iris_deletions;
+pub(crate) use iris_modifications::read_iris_modifications;
+pub(crate) use iris_shares::read_iris_shares_batch;
+pub(crate) use node_config::{read_node_config, NODE_CONFIG_KIND_GENESIS};
