+++ iris-mpc-upgrade-hawk/tests/utils/mod.rs
@@ -3,7 +3,6 @@ pub mod defaults;
 pub mod errors;
 pub mod logger;
 pub mod pgres;
-pub mod resources;
 pub mod runner;
 
 pub use errors::TestError;
